feat: send over user in SendSMS Hook instead of UserID (#1551)
## What kind of change does this PR introduce?

We align convention with `SendEmail` and send over a user to avoid
having the user make an additional `getUser` call. Also allows access to
`app_metadata` and `user_metadata` which would be useful for
internationalization where you may want the locale of the user to
determine which template to send.

We also introduce a `PhoneData` struct through which we can introduce
any potential phone related fields. This struct currently lives under
the `hooks` package as there is no `phone` package currently and
introducing one might require a significant refactor. Importing it as as
is under `api` package would cause a circular dependency between `hooks`
and `api` packages.

---------

Co-authored-by: Stojan Dimitrovski <sdimitrovski@gmail.com>