feat: add SAML specific external URL config (#1599)
Adds a SAML-specific external URL config, which allows the advertised
SAML metadata to be different than the one defined with the API external
URL.

This is useful in projects that use proxies or custom domains which can
be very disruptive with SAML as a new connection with the IDP needs to
be established. By configuring `GOTRUE_SAML_EXTERNAL_URL` to the URL
before the custom domain was set up, Auth will advertise the correct
metadata.