fix: remove check for content-length (#1700)
## What kind of change does this PR introduce?
* Not all responses are gonna contain the `Content-Length` header.
According to the HTTP/2 spec, it is not required to return the
content-length header in the response. If the "Transfer Encoding" is
chunked, the content-length header will also not be present as the
response is sent in chunks and the content-length is unknown initially.

## What is the current behavior?

Please link any relevant issues here.

## What is the new behavior?

Feel free to include screenshots if it includes visual changes.

## Additional context

Add any other context or screenshots.