fix: expose factor type on challenge (#1709)
## What kind of change does this PR introduce?

Expose factor type on challenge so that developers can know whether it
is a phone or TOTP factor they have challenged and redirect to
appropriate page or perform relevant action

Also helps us distinguish between the two so we can prevent unintended
behaviour (e.g. the use of `challengeAndVerify` with a phone factor)


As typescript follow structural typing I think addition of a new field
should be fine. We will test this before proceeding. However, putting it
out there first for early review or consideration

Update - tested and should be fine